Added the norecord and about files. Renamed my db.js file. Added more information to my index and about ejs files
