+++ views/about.ejs
@@ -1,0 +1,13 @@
+<html lang="en">
+    <%- include('./partials/head.ejs') %>
+    <body>
+        <%- include('./partials/nav.ejs') %>
+        <h1>About Us</h1>
+        <h2>Welcome to the Video Game Collective Database.</h2>
+        <br>
+        <p>This site was created by Samantha Thorne, a software development student at Keyin College. I'm very passionate about video games, and so I wanted to keep a record of the games I've played. </p>
+        <p>While you're here, feel free to explore, edit or delete from list of games. If you'd like to, you can even add your own favorites to the list. Who knows, I might even try out your suggestions!</p>
+        <br>
+    </body>
+    <%- include('./partials/foot.ejs') %>
+</html>
