+++ views/index.ejs
@@ -4,6 +4,11 @@
         <%- include('./partials/nav.ejs') %>
         <h1>Index Page</h1>
         <h2>Welcome to the index page.</h2>
+        <br>
+        <p>To view our list of video games, please click on the games link in the navigation bar.</p>
+        <p>To learn more about us, please click on the about link in the navigation bar.</p>
+        <br>
+        <br>
     </body>
     <%- include('./partials/foot.ejs') %>
 </html>
