+++ views/norecord.ejs
@@ -1,0 +1,10 @@
+<html lang="en">
+    <%- include('./partials/head.ejs') %>
+    <body>
+        <%- include('./partials/nav.ejs') %>
+        <div class="nodata content">
+            <h2>No data found for the route.</h2>
+        </div>
+    </body>
+    <%- include('./partials/foot.ejs') %>
+</html>
